Refactor files that are related to GUI to the gui package.
By abstracting out all the related code to the gui package, the code becomes more organised.  All the javafx-related files can be found in the gui package.
